+++ pganonymizer/cli.py
@@ -4,14 +4,18 @@ import time
 
 import yaml
 
+from pganonymizer.constants import DEFAULT_SCHEMA_FILE
+from pganonymizer.providers import PROVIDERS
 from pganonymizer.utils import anonymize_tables, get_connection, truncate_tables
 
 
 def main():
     parser = argparse.ArgumentParser(description='Anonymize data of a PostgreSQL database')
     parser.add_argument('-v', '--verbose', action='count', help='Increase verbosity')
-    parser.add_argument('--schema', help='A YAML file that contains the anonymization rules', required=True,
-                        default='./schema.yml')
+    parser.add_argument('--list-providers', action='store_true', help='Show a list of all available providers',
+                        default=False)
+    parser.add_argument('--schema', help='A YAML file that contains the anonymization rules',
+                        default=DEFAULT_SCHEMA_FILE)
     parser.add_argument('--dbname', help='Name of the database')
     parser.add_argument('--user', help='Name of the database user')
     parser.add_argument('--password', default='', help='Password for the database user')
@@ -20,10 +24,16 @@ def main():
     parser.add_argument('--dry-run', action='store_true', help='Don\'t commit changes made on the database',
                         default=False)
     args = parser.parse_args()
+
     loglevel = logging.WARNING
     if args.verbose:
         loglevel = logging.DEBUG
     logging.basicConfig(format='%(levelname)s: %(message)s', level=loglevel)
+
+    if args.list_providers:
+        print('Available provider classes:\n')
+        for provider_cls in PROVIDERS:
+            print('{:<10} {}'.format(provider_cls.id, provider_cls.__doc__))
 
     schema = yaml.load(open(args.schema), Loader=yaml.FullLoader)
     connection = get_connection(args)